move handlePodResourcesResize unit tests into allocation_manager_test
