+++ pkg/kubelet/kubelet_test.go
@@ -2633,605 +2633,6 @@ func TestPodResourceAllocationReset(t *testing.T) {
 	}
 }
 
-func TestHandlePodResourcesResizeWithSwap(t *testing.T) {
-	if goruntime.GOOS == "windows" {
-		t.Skip("InPlacePodVerticalScaling is not currently supported for Windows")
-	}
-	featuregatetesting.SetFeatureGateDuringTest(t, utilfeature.DefaultFeatureGate, features.InPlacePodVerticalScaling, true)
-	featuregatetesting.SetFeatureGateDuringTest(t, utilfeature.DefaultFeatureGate, features.NodeSwap, true)
-	testKubelet := newTestKubelet(t, false)
-	defer testKubelet.Cleanup()
-	noSwapContainerName, swapContainerName := "test-container-noswap", "test-container-limitedswap"
-	runtime := testKubelet.fakeRuntime
-	runtime.SwapBehavior = map[string]kubetypes.SwapBehavior{
-		noSwapContainerName: kubetypes.NoSwap,
-		swapContainerName:   kubetypes.LimitedSwap,
-	}
-	kubelet := testKubelet.kubelet
-	cpu500m := resource.MustParse("500m")
-	cpu1000m := resource.MustParse("1")
-	mem500M := resource.MustParse("500Mi")
-	mem1000M := resource.MustParse("1Gi")
-	nodes := []*v1.Node{
-		{
-			ObjectMeta: metav1.ObjectMeta{Name: testKubeletHostname},
-			Status: v1.NodeStatus{
-				Capacity: v1.ResourceList{
-					v1.ResourceCPU:    resource.MustParse("8"),
-					v1.ResourceMemory: resource.MustParse("8Gi"),
-				},
-				Allocatable: v1.ResourceList{
-					v1.ResourceCPU:    resource.MustParse("4"),
-					v1.ResourceMemory: resource.MustParse("4Gi"),
-					v1.ResourcePods:   *resource.NewQuantity(40, resource.DecimalSI),
-				},
-			},
-		},
-	}
-	kubelet.nodeLister = testNodeLister{nodes: nodes}
-	testPod := &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			UID:       "1111",
-			Name:      "pod1",
-			Namespace: "ns1",
-		},
-		Spec: v1.PodSpec{
-			Containers: []v1.Container{
-				{
-					Name:  "c1",
-					Image: "i1",
-					Resources: v1.ResourceRequirements{
-						Requests: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-					},
-				},
-			},
-		},
-		Status: v1.PodStatus{
-			Phase: v1.PodRunning,
-			ContainerStatuses: []v1.ContainerStatus{
-				{
-					Name:               "c1",
-					AllocatedResources: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-					Resources:          &v1.ResourceRequirements{},
-				},
-			},
-		},
-	}
-
-	testKubelet.fakeKubeClient = fake.NewSimpleClientset(testPod)
-	kubelet.kubeClient = testKubelet.fakeKubeClient
-	defer testKubelet.fakeKubeClient.ClearActions()
-	kubelet.podManager.AddPod(testPod)
-	kubelet.podWorkers.(*fakePodWorkers).running = map[types.UID]bool{
-		testPod.UID: true,
-	}
-	defer kubelet.podManager.RemovePod(testPod)
-	tests := []struct {
-		name                  string
-		newRequests           v1.ResourceList
-		expectedAllocatedReqs v1.ResourceList
-		resizePolicy          v1.ContainerResizePolicy
-		swapBehavior          kubetypes.SwapBehavior
-		expectedResize        []*v1.PodCondition
-	}{
-		{
-			name:                  "NoSwap Request Memory decrease ResizePolicy RestartContainer - expect InProgress",
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem500M},
-			swapBehavior:          kubetypes.NoSwap,
-			resizePolicy:          v1.ContainerResizePolicy{ResourceName: v1.ResourceMemory, RestartPolicy: v1.RestartContainer},
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-		{
-			name:                  "LimitedSwap Request Memory increase with ResizePolicy RestartContainer - expect InProgress",
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem500M},
-			swapBehavior:          kubetypes.LimitedSwap,
-			resizePolicy:          v1.ContainerResizePolicy{ResourceName: v1.ResourceMemory, RestartPolicy: v1.RestartContainer},
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-		{
-			name:                  "LimitedSwap Request Memory increase with ResizePolicy NotRequired - expect Infeasible",
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			swapBehavior:          kubetypes.LimitedSwap,
-			resizePolicy:          v1.ContainerResizePolicy{ResourceName: v1.ResourceMemory, RestartPolicy: v1.NotRequired},
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:    v1.PodResizePending,
-					Status:  "True",
-					Reason:  "Infeasible",
-					Message: "In-place resize of containers with swap is not supported",
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			originalPod := testPod.DeepCopy()
-			originalPod.Spec.Containers[0].ResizePolicy = []v1.ContainerResizePolicy{tt.resizePolicy}
-			if tt.swapBehavior == kubetypes.NoSwap {
-				originalPod.Spec.Containers[0].Name = noSwapContainerName
-			} else {
-				originalPod.Spec.Containers[0].Name = swapContainerName
-			}
-			kubelet.podManager.UpdatePod(originalPod)
-			newPod := originalPod.DeepCopy()
-			newPod.Spec.Containers[0].Resources.Requests = tt.newRequests
-			require.NoError(t, kubelet.allocationManager.SetAllocatedResources(originalPod))
-			require.NoError(t, kubelet.allocationManager.SetActuatedResources(originalPod, nil))
-			t.Cleanup(func() { kubelet.allocationManager.RemovePod(originalPod.UID) })
-			podStatus := &kubecontainer.PodStatus{
-				ID:        originalPod.UID,
-				Name:      originalPod.Name,
-				Namespace: originalPod.Namespace,
-			}
-			setContainerStatus := func(podStatus *kubecontainer.PodStatus, c *v1.Container, idx int) {
-				podStatus.ContainerStatuses[idx] = &kubecontainer.Status{
-					Name:  c.Name,
-					State: kubecontainer.ContainerStateRunning,
-					Resources: &kubecontainer.ContainerResources{
-						CPURequest:  c.Resources.Requests.Cpu(),
-						CPULimit:    c.Resources.Limits.Cpu(),
-						MemoryLimit: c.Resources.Limits.Memory(),
-					},
-				}
-			}
-			podStatus.ContainerStatuses = make([]*kubecontainer.Status, len(originalPod.Spec.Containers))
-			for i, c := range originalPod.Spec.Containers {
-				setContainerStatus(podStatus, &c, i)
-			}
-
-			updatedPod, err := kubelet.allocationManager.HandlePodResourcesResize(kubelet.containerRuntime, kubelet.getAllocatedPods(), newPod, podStatus)
-			require.NoError(t, err)
-
-			updatedPodCtr := updatedPod.Spec.Containers[0]
-			assert.Equal(t, tt.expectedAllocatedReqs, updatedPodCtr.Resources.Requests, "updated pod spec requests")
-
-			alloc, found := kubelet.allocationManager.GetContainerResourceAllocation(newPod.UID, updatedPodCtr.Name)
-			require.True(t, found, "container allocation")
-			assert.Equal(t, tt.expectedAllocatedReqs, alloc.Requests, "stored container request allocation")
-
-			resizeStatus := kubelet.statusManager.GetPodResizeConditions(newPod.UID)
-			for i := range resizeStatus {
-				// Ignore probe time and last transition time during comparison.
-				resizeStatus[i].LastProbeTime = metav1.Time{}
-				resizeStatus[i].LastTransitionTime = metav1.Time{}
-				assert.Contains(t, resizeStatus[i].Message, tt.expectedResize[i].Message)
-				resizeStatus[i].Message = tt.expectedResize[i].Message
-			}
-			assert.Equal(t, tt.expectedResize, resizeStatus)
-		})
-	}
-}
-
-func TestHandlePodResourcesResize(t *testing.T) {
-	if goruntime.GOOS == "windows" {
-		t.Skip("InPlacePodVerticalScaling is not currently supported for Windows")
-	}
-	featuregatetesting.SetFeatureGateDuringTest(t, utilfeature.DefaultFeatureGate, features.InPlacePodVerticalScaling, true)
-	testKubelet := newTestKubelet(t, false)
-	defer testKubelet.Cleanup()
-	kubelet := testKubelet.kubelet
-	containerRestartPolicyAlways := v1.ContainerRestartPolicyAlways
-
-	cpu2m := resource.MustParse("2m")
-	cpu500m := resource.MustParse("500m")
-	cpu1000m := resource.MustParse("1")
-	cpu1500m := resource.MustParse("1500m")
-	cpu2500m := resource.MustParse("2500m")
-	cpu5000m := resource.MustParse("5000m")
-	mem500M := resource.MustParse("500Mi")
-	mem1000M := resource.MustParse("1Gi")
-	mem1500M := resource.MustParse("1500Mi")
-	mem2500M := resource.MustParse("2500Mi")
-	mem4500M := resource.MustParse("4500Mi")
-
-	nodes := []*v1.Node{
-		{
-			ObjectMeta: metav1.ObjectMeta{Name: testKubeletHostname},
-			Status: v1.NodeStatus{
-				Capacity: v1.ResourceList{
-					v1.ResourceCPU:    resource.MustParse("8"),
-					v1.ResourceMemory: resource.MustParse("8Gi"),
-				},
-				Allocatable: v1.ResourceList{
-					v1.ResourceCPU:    resource.MustParse("4"),
-					v1.ResourceMemory: resource.MustParse("4Gi"),
-					v1.ResourcePods:   *resource.NewQuantity(40, resource.DecimalSI),
-				},
-			},
-		},
-	}
-	kubelet.nodeLister = testNodeLister{nodes: nodes}
-
-	testPod1 := &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			UID:       "1111",
-			Name:      "pod1",
-			Namespace: "ns1",
-		},
-		Spec: v1.PodSpec{
-			Containers: []v1.Container{
-				{
-					Name:  "c1",
-					Image: "i1",
-					Resources: v1.ResourceRequirements{
-						Requests: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-					},
-				},
-			},
-		},
-		Status: v1.PodStatus{
-			Phase: v1.PodRunning,
-			ContainerStatuses: []v1.ContainerStatus{
-				{
-					Name:               "c1",
-					AllocatedResources: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-					Resources:          &v1.ResourceRequirements{},
-				},
-			},
-		},
-	}
-	testPod2 := testPod1.DeepCopy()
-	testPod2.UID = "2222"
-	testPod2.Name = "pod2"
-	testPod2.Namespace = "ns2"
-	testPod2.Spec = v1.PodSpec{
-		InitContainers: []v1.Container{
-			{
-				Name:  "c1-init",
-				Image: "i1",
-				Resources: v1.ResourceRequirements{
-					Requests: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-				},
-				RestartPolicy: &containerRestartPolicyAlways,
-			},
-		},
-	}
-	testPod2.Status = v1.PodStatus{
-		Phase: v1.PodRunning,
-		InitContainerStatuses: []v1.ContainerStatus{
-			{
-				Name:               "c1-init",
-				AllocatedResources: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-				Resources:          &v1.ResourceRequirements{},
-			},
-		},
-	}
-	testPod3 := testPod1.DeepCopy()
-	testPod3.UID = "3333"
-	testPod3.Name = "pod3"
-	testPod3.Namespace = "ns2"
-
-	testKubelet.fakeKubeClient = fake.NewSimpleClientset(testPod1, testPod2, testPod3)
-	kubelet.kubeClient = testKubelet.fakeKubeClient
-	defer testKubelet.fakeKubeClient.ClearActions()
-	kubelet.podManager.AddPod(testPod1)
-	kubelet.podManager.AddPod(testPod2)
-	kubelet.podManager.AddPod(testPod3)
-	kubelet.podWorkers.(*fakePodWorkers).running = map[types.UID]bool{
-		testPod1.UID: true,
-		testPod2.UID: true,
-		testPod3.UID: true,
-	}
-	defer kubelet.podManager.RemovePod(testPod3)
-	defer kubelet.podManager.RemovePod(testPod2)
-	defer kubelet.podManager.RemovePod(testPod1)
-
-	tests := []struct {
-		name                  string
-		originalRequests      v1.ResourceList
-		newRequests           v1.ResourceList
-		originalLimits        v1.ResourceList
-		newLimits             v1.ResourceList
-		newResourcesAllocated bool // Whether the new requests have already been allocated (but not actuated)
-		expectedAllocatedReqs v1.ResourceList
-		expectedAllocatedLims v1.ResourceList
-		expectedResize        []*v1.PodCondition
-		annotations           map[string]string
-	}{
-		{
-			name:                  "Request CPU and memory decrease - expect InProgress",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem500M},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-		{
-			name:                  "Request CPU increase, memory decrease - expect InProgress",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu1500m, v1.ResourceMemory: mem500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1500m, v1.ResourceMemory: mem500M},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-		{
-			name:                  "Request CPU decrease, memory increase - expect InProgress",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem1500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem1500M},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-		{
-			name:                  "Request CPU and memory increase beyond current capacity - expect Deferred",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu2500m, v1.ResourceMemory: mem2500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:    v1.PodResizePending,
-					Status:  "True",
-					Reason:  "Deferred",
-					Message: "",
-				},
-			},
-		},
-		{
-			name:                  "Request CPU decrease and memory increase beyond current capacity - expect Deferred",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem2500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:    v1.PodResizePending,
-					Status:  "True",
-					Reason:  "Deferred",
-					Message: "Node didn't have enough resource: memory",
-				},
-			},
-		},
-		{
-			name:                  "Request memory increase beyond node capacity - expect Infeasible",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem4500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:    v1.PodResizePending,
-					Status:  "True",
-					Reason:  "Infeasible",
-					Message: "Node didn't have enough capacity: memory, requested: 4718592000, capacity: 4294967296",
-				},
-			},
-		},
-		{
-			name:                  "Request CPU increase beyond node capacity - expect Infeasible",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu5000m, v1.ResourceMemory: mem1000M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:    v1.PodResizePending,
-					Status:  "True",
-					Reason:  "Infeasible",
-					Message: "Node didn't have enough capacity: cpu, requested: 5000, capacity: 4000",
-				},
-			},
-		},
-		{
-			name:                  "CPU increase in progress - expect InProgress",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu1500m, v1.ResourceMemory: mem1000M},
-			newResourcesAllocated: true,
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1500m, v1.ResourceMemory: mem1000M},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-		{
-			name:                  "No resize",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			expectedResize:        nil,
-		},
-		{
-			name:                  "static pod, expect Infeasible",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu500m, v1.ResourceMemory: mem500M},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1000m, v1.ResourceMemory: mem1000M},
-			annotations:           map[string]string{kubetypes.ConfigSourceAnnotationKey: kubetypes.FileSource},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:    v1.PodResizePending,
-					Status:  "True",
-					Reason:  "Infeasible",
-					Message: "In-place resize of static-pods is not supported",
-				},
-			},
-		},
-		{
-			name:                  "Increase CPU from min shares",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu2m},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu1000m},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu1000m},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-		{
-			name:                  "Decrease CPU to min shares",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: cpu1000m},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: cpu2m},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: cpu2m},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-		{
-			name:                  "Increase CPU from min limit",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: resource.MustParse("10m")},
-			originalLimits:        v1.ResourceList{v1.ResourceCPU: resource.MustParse("10m")},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: resource.MustParse("10m")}, // Unchanged
-			newLimits:             v1.ResourceList{v1.ResourceCPU: resource.MustParse("20m")},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: resource.MustParse("10m")},
-			expectedAllocatedLims: v1.ResourceList{v1.ResourceCPU: resource.MustParse("20m")},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-		{
-			name:                  "Decrease CPU to min limit",
-			originalRequests:      v1.ResourceList{v1.ResourceCPU: resource.MustParse("10m")},
-			originalLimits:        v1.ResourceList{v1.ResourceCPU: resource.MustParse("20m")},
-			newRequests:           v1.ResourceList{v1.ResourceCPU: resource.MustParse("10m")}, // Unchanged
-			newLimits:             v1.ResourceList{v1.ResourceCPU: resource.MustParse("10m")},
-			expectedAllocatedReqs: v1.ResourceList{v1.ResourceCPU: resource.MustParse("10m")},
-			expectedAllocatedLims: v1.ResourceList{v1.ResourceCPU: resource.MustParse("10m")},
-
-			expectedResize: []*v1.PodCondition{
-				{
-					Type:   v1.PodResizeInProgress,
-					Status: "True",
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		for _, isSidecarContainer := range []bool{false, true} {
-			t.Run(fmt.Sprintf("%s/sidecar=%t", tt.name, isSidecarContainer), func(t *testing.T) {
-				var originalPod *v1.Pod
-				var originalCtr *v1.Container
-				if isSidecarContainer {
-					originalPod = testPod2.DeepCopy()
-					originalCtr = &originalPod.Spec.InitContainers[0]
-				} else {
-					originalPod = testPod1.DeepCopy()
-					originalCtr = &originalPod.Spec.Containers[0]
-				}
-				originalPod.Annotations = tt.annotations
-				originalCtr.Resources.Requests = tt.originalRequests
-				originalCtr.Resources.Limits = tt.originalLimits
-
-				kubelet.podManager.UpdatePod(originalPod)
-
-				newPod := originalPod.DeepCopy()
-
-				if isSidecarContainer {
-					newPod.Spec.InitContainers[0].Resources.Requests = tt.newRequests
-					newPod.Spec.InitContainers[0].Resources.Limits = tt.newLimits
-				} else {
-					newPod.Spec.Containers[0].Resources.Requests = tt.newRequests
-					newPod.Spec.Containers[0].Resources.Limits = tt.newLimits
-				}
-
-				if !tt.newResourcesAllocated {
-					require.NoError(t, kubelet.allocationManager.SetAllocatedResources(originalPod))
-				} else {
-					require.NoError(t, kubelet.allocationManager.SetAllocatedResources(newPod))
-				}
-				require.NoError(t, kubelet.allocationManager.SetActuatedResources(originalPod, nil))
-				t.Cleanup(func() { kubelet.allocationManager.RemovePod(originalPod.UID) })
-
-				podStatus := &kubecontainer.PodStatus{
-					ID:        originalPod.UID,
-					Name:      originalPod.Name,
-					Namespace: originalPod.Namespace,
-				}
-
-				setContainerStatus := func(podStatus *kubecontainer.PodStatus, c *v1.Container, idx int) {
-					podStatus.ContainerStatuses[idx] = &kubecontainer.Status{
-						Name:  c.Name,
-						State: kubecontainer.ContainerStateRunning,
-						Resources: &kubecontainer.ContainerResources{
-							CPURequest:  c.Resources.Requests.Cpu(),
-							CPULimit:    c.Resources.Limits.Cpu(),
-							MemoryLimit: c.Resources.Limits.Memory(),
-						},
-					}
-				}
-
-				podStatus.ContainerStatuses = make([]*kubecontainer.Status, len(originalPod.Spec.Containers)+len(originalPod.Spec.InitContainers))
-				for i, c := range originalPod.Spec.InitContainers {
-					setContainerStatus(podStatus, &c, i)
-				}
-				for i, c := range originalPod.Spec.Containers {
-					setContainerStatus(podStatus, &c, i+len(originalPod.Spec.InitContainers))
-				}
-
-				updatedPod, err := kubelet.allocationManager.HandlePodResourcesResize(kubelet.containerRuntime, kubelet.getAllocatedPods(), newPod, podStatus)
-				require.NoError(t, err)
-
-				var updatedPodCtr v1.Container
-				if isSidecarContainer {
-					updatedPodCtr = updatedPod.Spec.InitContainers[0]
-				} else {
-					updatedPodCtr = updatedPod.Spec.Containers[0]
-				}
-				assert.Equal(t, tt.expectedAllocatedReqs, updatedPodCtr.Resources.Requests, "updated pod spec requests")
-				assert.Equal(t, tt.expectedAllocatedLims, updatedPodCtr.Resources.Limits, "updated pod spec limits")
-
-				alloc, found := kubelet.allocationManager.GetContainerResourceAllocation(newPod.UID, updatedPodCtr.Name)
-				require.True(t, found, "container allocation")
-				assert.Equal(t, tt.expectedAllocatedReqs, alloc.Requests, "stored container request allocation")
-				assert.Equal(t, tt.expectedAllocatedLims, alloc.Limits, "stored container limit allocation")
-
-				resizeStatus := kubelet.statusManager.GetPodResizeConditions(newPod.UID)
-				for i := range resizeStatus {
-					// Ignore probe time and last transition time during comparison.
-					resizeStatus[i].LastProbeTime = metav1.Time{}
-					resizeStatus[i].LastTransitionTime = metav1.Time{}
-
-					// Message is a substring assertion, since it can change slightly.
-					assert.Contains(t, resizeStatus[i].Message, tt.expectedResize[i].Message)
-					resizeStatus[i].Message = tt.expectedResize[i].Message
-				}
-				assert.Equal(t, tt.expectedResize, resizeStatus)
-			})
-		}
-	}
-}
-
 // testPodSyncLoopHandler is a lifecycle.PodSyncLoopHandler that is used for testing.
 type testPodSyncLoopHandler struct {
 	// list of pods to sync
